requirements.txt
anyio==3.3.2
APScheduler==3.6.3
asgiref==3.4.1
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
google-api-core==2.0.1
google-auth==2.2.1
google-cloud-core==2.0.0
google-cloud-firestore==2.3.4
googleapis-common-protos==1.53.0
grpcio==1.41.0
h11==0.12.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
proto-plus==1.19.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-multipart==0.0.5
python-telegram-bot==13.7
pytz==2021.3
regex==2021.9.30
requests==2.26.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.1
tornado==6.1
typing-extensions==3.10.0.2
tzlocal==3.0
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
